Trait rusty_nail::dispenser::Dispenser

pub trait Dispenser {
    fn max_flow_rate(&self) -> f64;
    fn set_flow_rate(&mut self, rate: f64) -> Result<(), RustyNailError>;
    fn get_name(&self) -> Result<&String, RustyNailError>;
    fn set_name(&mut self, str: String);
    fn set_level(&mut self, level: f64) -> Result<(), RustyNailError>;
    fn get_level(&self) -> Result<f64, RustyNailError>;
    fn dispense(&mut self, quantity: f64) -> Result<f64, RustyNailError>;
}

Defines trait for interfacing to a dispense

Required Methods

fn max_flow_rate(&self) -> f64

Returns the maximum flow rate of the pump in mL/s

fn set_flow_rate(&mut self, rate: f64) -> Result<(), RustyNailError>

Sets the current flow rate of the pump in mL/s

fn get_name(&self) -> Result<&String, RustyNailError>

Get the name of beverage being pumped.

fn set_name(&mut self, str: String)

Get the name of beverage being pumped.

fn set_level(&mut self, level: f64) -> Result<(), RustyNailError>

Set the current liquid level in mL.

fn get_level(&self) -> Result<f64, RustyNailError>

Get the current liquid level in mL.

fn dispense(&mut self, quantity: f64) -> Result<f64, RustyNailError>

Dispense some quantity of liquid in mL.